Move OrleansException to another file and move it to Abstractions
